Recipe Class

This class represents a recipe for processing survey data. A recipe includes metadata such as name, edition, survey type, user, description, and a list of steps to be applied to the survey data.

Public fields

name

A string representing the name of the recipe.

edition

A string indicating the edition of the survey associated with the recipe.

survey_type

A string specifying the type of survey (e.g., "eaii", "ech").

default_engine

A string indicating the default engine used for processing.

depends_on

A list of dependencies required by the recipe.

user

A string representing the user or creator of the recipe.

description

A string providing a description of the recipe.

id

A unique identifier for the recipe.

steps

A list of steps to be applied to the survey data.

doi

A string representing the DOI (Digital Object Identifier) of the recipe.

bake

A logical value indicating whether the recipe should be baked (processed) immediately.

topic

A string specifying the topic of the recipe.

Methods

Public methods

Recipe$new()

Recipe$clone()

Method new()

Initializes a new instance of the Recipe class.

Usage

Recipe$new(
  name,
  edition,
  survey_type,
  default_engine,
  depends_on,
  user,
  description,
  steps,
  id,
  doi,
  topic
)

Arguments

name

A string representing the name of the recipe.

edition

A string indicating the edition of the survey associated with the recipe.

survey_type

A string specifying the type of survey (e.g., "eaii", "ech").

default_engine

A string indicating the default engine used for processing.

depends_on

A list of dependencies required by the recipe.

user

A string representing the user or creator of the recipe.

description

A string providing a description of the recipe.

steps

A list of steps to be applied to the survey data.

id

A unique identifier for the recipe.

doi

A string representing the DOI (Digital Object Identifier) of the recipe.

topic

A string specifying the topic of the recipe.

Method clone()

The objects of this class are cloneable with this method.

Usage

Recipe$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.