Recipe

Creates a new Recipe object for processing survey data.

Usage

recipe(...)

Arguments

...

A list with the following metadata: name, user, svy, description.

Value

A Recipe object.

Details

This function validates the provided metadata and creates a Recipe object. If steps are included, they are added to the Recipe object.

See also

Examples

# Example of creating a Recipe object
recipe_obj <- recipe(
  name = "Example Recipe",
  user = "Metasurvey User",
  svy = survey_empty(type = "eaii", edition = "2019-2021"),
  description = "This is an example recipe."
)
print(recipe_obj)
#> <Recipe>
#>   Public:
#>     bake: FALSE
#>     clone: function (deep = FALSE) 
#>     default_engine: data.table
#>     depends_on: list
#>     description: This is an example recipe.
#>     doi: NULL
#>     edition: 2019_2021
#>     id: 0.289892295375466
#>     initialize: function (name, edition, survey_type, default_engine, depends_on, 
#>     name: Example Recipe
#>     steps: list
#>     survey_type: eaii
#>     topic: NULL
#>     user: Metasurvey User