metasurvey: Survey Processing with Meta-Programming

Provides a set of functions to process survey data using meta-programming. It is designed to work with the 'survey' package, but it can be used with any other package that provides a similar data structure. The main goal of this package is to provide a set of functions that can be used to process survey data in a more efficient way, by using meta-programming to generate code that can be used to process the data.

Author

Maintainer: Mauro Loprete mloprete@cognus.cl

Authors:

Natalia da Silva